+++ server/runtime_nakama_module.go
@@ -16,6 +16,9 @@ package server
 
 import (
 	"context"
+	"io"
+	"net/http"
+	"time"
 
 	"strings"
 
@@ -26,6 +29,9 @@ import (
 	"encoding/json"
 
 	"encoding/base64"
+
+	"encoding/hex"
+	"io/ioutil"
 
 	"github.com/fatih/structs"
 	"github.com/satori/go.uuid"
@@ -46,6 +52,7 @@ type NakamaModule struct {
 	logger              *zap.Logger
 	db                  *sql.DB
 	notificationService *NotificationService
+	client              *http.Client
 }
 
 func NewNakamaModule(logger *zap.Logger, db *sql.DB, l *lua.LState, notificationService *NotificationService) *NakamaModule {
@@ -59,11 +66,22 @@ func NewNakamaModule(logger *zap.Logger, db *sql.DB, l *lua.LState, notification
 		logger:              logger,
 		db:                  db,
 		notificationService: notificationService,
+		client: &http.Client{
+			Timeout: 5 * time.Second,
+		},
 	}
 }
 
 func (n *NakamaModule) Loader(l *lua.LState) int {
 	mod := l.SetFuncs(l.NewTable(), map[string]lua.LGFunction{
+		"uuid_v4":               n.uuidV4,
+		"http_request":          n.httpRequest,
+		"json_encode":           n.jsonEncode,
+		"json_decode":           n.jsonDecode,
+		"base64_encode":         n.base64Encode,
+		"base64_decode":         n.base64Decode,
+		"base16_encode":         n.base16Encode,
+		"base16_decode":         n.base16decode,
 		"logger_info":           n.loggerInfo,
 		"logger_warn":           n.loggerWarn,
 		"logger_error":          n.loggerError,
@@ -85,6 +103,166 @@ func (n *NakamaModule) Loader(l *lua.LState) int {
 	})
 
 	l.Push(mod)
+	return 1
+}
+
+func (n *NakamaModule) uuidV4(l *lua.LState) int {
+	// TODO ensure there were no arguments to the function
+	l.Push(lua.LString(uuid.NewV4().String()))
+	return 1
+}
+
+func (n *NakamaModule) httpRequest(l *lua.LState) int {
+	url := l.CheckString(1)
+	method := l.CheckString(2)
+	headers := l.CheckTable(3)
+	body := l.OptString(4, "")
+	if url == "" {
+		l.ArgError(1, "Expects URL string")
+		return 0
+	}
+	if method == "" {
+		l.ArgError(2, "Expects method string")
+		return 0
+	}
+
+	// Prepare request body, if any.
+	var requestBody io.Reader
+	if body != "" {
+		requestBody = strings.NewReader(body)
+	}
+	// Prepare the request.
+	req, err := http.NewRequest(method, url, requestBody)
+	if err != nil {
+		l.RaiseError("HTTP request error: %v", err.Error())
+		return 0
+	}
+	// Apply any request headers.
+	httpHeaders := ConvertLuaTable(headers)
+	for k, v := range httpHeaders {
+		if vs, ok := v.(string); !ok {
+			l.RaiseError("HTTP header values must be strings")
+			return 0
+		} else {
+			req.Header.Add(k, vs)
+		}
+	}
+	// Execute the request.
+	resp, err := n.client.Do(req)
+	if err != nil {
+		l.RaiseError("HTTP request error: %v", err.Error())
+		return 0
+	}
+	// Read the response body.
+	responseBody, err := ioutil.ReadAll(resp.Body)
+	resp.Body.Close()
+	if err != nil {
+		l.RaiseError("HTTP response body error: %v", err.Error())
+		return 0
+	}
+	// Read the response headers.
+	responseHeaders := make(map[string]interface{}, len(resp.Header))
+	for k, vs := range resp.Header {
+		// TODO accept multiple values per header
+		for _, v := range vs {
+			responseHeaders[k] = v
+			break
+		}
+	}
+
+	l.Push(lua.LNumber(resp.StatusCode))
+	l.Push(ConvertMap(l, responseHeaders))
+	l.Push(lua.LString(string(responseBody)))
+	return 3
+}
+
+func (n *NakamaModule) jsonEncode(l *lua.LState) int {
+	jsonTable := l.Get(1)
+	if jsonTable == nil {
+		l.ArgError(1, "Expects a non-nil value to encode")
+		return 0
+	}
+
+	jsonData := convertLuaValue(jsonTable)
+	jsonBytes, err := json.Marshal(jsonData)
+	if err != nil {
+		l.RaiseError("Error encoding to JSON: %v", err.Error())
+		return 0
+	}
+
+	l.Push(lua.LString(string(jsonBytes)))
+	return 1
+}
+
+func (n *NakamaModule) jsonDecode(l *lua.LState) int {
+	jsonString := l.CheckString(1)
+	if jsonString == "" {
+		l.ArgError(1, "Expects JSON string")
+		return 0
+	}
+
+	var jsonData interface{}
+	if err := json.Unmarshal([]byte(jsonString), &jsonData); err != nil {
+		l.RaiseError("Not a valid JSON string: %v", err.Error())
+		return 0
+	}
+
+	l.Push(convertValue(l, jsonData))
+	return 1
+}
+
+func (n *NakamaModule) base64Encode(l *lua.LState) int {
+	input := l.CheckString(1)
+	if input == "" {
+		l.ArgError(1, "Expects string")
+		return 0
+	}
+
+	output := base64.StdEncoding.EncodeToString([]byte(input))
+	l.Push(lua.LString(output))
+	return 1
+}
+func (n *NakamaModule) base64Decode(l *lua.LState) int {
+	input := l.CheckString(1)
+	if input == "" {
+		l.ArgError(1, "Expects string")
+		return 0
+	}
+
+	output, err := base64.StdEncoding.DecodeString(input)
+	if err != nil {
+		l.RaiseError("Not a valid base64 string: %v", err.Error())
+		return 0
+	}
+
+	l.Push(lua.LString(output))
+	return 1
+}
+func (n *NakamaModule) base16Encode(l *lua.LState) int {
+	input := l.CheckString(1)
+	if input == "" {
+		l.ArgError(1, "Expects string")
+		return 0
+	}
+
+	output := hex.EncodeToString([]byte(input))
+	l.Push(lua.LString(output))
+	return 1
+}
+func (n *NakamaModule) base16decode(l *lua.LState) int {
+	input := l.CheckString(1)
+	if input == "" {
+		l.ArgError(1, "Expects string")
+		return 0
+	}
+
+	output, err := hex.DecodeString(input)
+	if err != nil {
+		l.RaiseError("Not a valid base16 string: %v", err.Error())
+		return 0
+	}
+
+	l.Push(lua.LString(output))
 	return 1
 }
 
@@ -361,11 +539,19 @@ func (n *NakamaModule) usersUpdate(l *lua.LState) int {
 				}
 				update.Lang = v.String()
 			case "Metadata":
-				if v.Type() != lua.LTString {
-					conversionError = "expects valid metadata in each update"
+				if v.Type() != lua.LTTable {
+					conversionError = "expects Metadata to be a table"
 					return
 				}
-				update.Metadata = []byte(v.String())
+
+				metadataMap := ConvertLuaTable(v.(*lua.LTable))
+				metadataBytes, err := json.Marshal(metadataMap)
+				if err != nil {
+					conversionError = fmt.Sprintf("failed to convert metadata: %s", err.Error())
+					return
+				}
+
+				update.Metadata = metadataBytes
 			case "AvatarUrl":
 				if v.Type() != lua.LTString {
 					conversionError = "expects valid avatar urls in each update"
@@ -943,20 +1129,21 @@ func (n *NakamaModule) groupsCreate(l *lua.LState) int {
 				}
 				p.Private = lua.LVAsBool(v)
 			case "Metadata":
-				if v.Type() != lua.LTString {
+				if v.Type() != lua.LTTable {
 					conversionError = true
-					l.ArgError(1, "expects Metadata to be string")
+					l.ArgError(1, "expects Metadata to be a table")
 					return
 				}
 
-				j := []byte(v.String())
-				maybeJson := make(map[string]interface{})
-				if jsonErr := json.Unmarshal(j, &maybeJson); jsonErr != nil {
+				metadataMap := ConvertLuaTable(v.(*lua.LTable))
+				metadataBytes, err := json.Marshal(metadataMap)
+				if err != nil {
 					conversionError = true
-					l.ArgError(1, "expects Metadata to be JSON")
+					l.ArgError(1, fmt.Sprintf("failed to convert metadata: %s", err.Error()))
 					return
 				}
-				p.Metadata = j
+
+				p.Metadata = metadataBytes
 			case "CreatorId":
 				if v.Type() != lua.LTString {
 					conversionError = true
@@ -1046,18 +1233,21 @@ func (n *NakamaModule) notificationsSendId(l *lua.LState) int {
 				}
 				notification.Subject = v.String()
 			case "Content":
-				if v.Type() != lua.LTString {
+				if v.Type() != lua.LTTable {
 					conversionError = true
-					l.ArgError(1, "expects Content to be string")
+					l.ArgError(1, "expects Content to be a table")
 					return
 				}
-				j := []byte(v.String())
-				maybeJson := make(map[string]interface{})
-				if jsonErr := json.Unmarshal(j, &maybeJson); jsonErr != nil {
-					l.ArgError(1, "expects Content to be a valid JSON")
+
+				contentMap := ConvertLuaTable(v.(*lua.LTable))
+				contentBytes, err := json.Marshal(contentMap)
+				if err != nil {
+					conversionError = true
+					l.ArgError(1, fmt.Sprintf("failed to convert content: %s", err.Error()))
 					return
 				}
-				notification.Content = j
+
+				notification.Content = contentBytes
 			case "Code":
 				if v.Type() != lua.LTNumber {
 					conversionError = true