+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com/) and this project uses [semantic versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Use Lua table for Content field when creating new notifications.
+- Use Lua table for Metadata field for new groups.
+- Use Lua table for Metadata field when updating a user.
+- Updated configuration variable names. Most importantly `DB` is now changed to `database.address`.
+- Moved all `nakamax` functions into `nakama`. 
 
 ## [1.0.0-rc.1] - 2017-07-18
 ### Added